[EDSD-29] feat: 단어수집 Service 구현 및 단위테스트 추가

+++ backend/clip-server/src/main/java/com/clip/server/word/dto/response/CollectedWordResponse.java
@@ -17,5 +17,5 @@ public class CollectedWordResponse {
     // 단어 수집 시간
     private LocalDateTime collectedAt;
     // 총 수집 단어
-    private int totalCollectedWords;
+    private Long totalCollectedWords;
 }
